requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
supabase
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-multipart